requirements.txt
fastapi==0.82.0  
pandas==1.4.2    
sqlalchemy==1.4.22
pyodbc==4.0.34       
uvicorn==0.16.0  
pydantic==1.8.2
python-dotenv==0.20.0  
selenium==4.21.0
lxml==5.2.2
alembic==1.13.1
